# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "保哥-前台-API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "31-review-develop-82qm1o.atavral.ninja",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "登入",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "登入",
                "parameters": [
                    {
                        "description": "登入資料",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/member.LoginRequestModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/member.LoginResponseModel"
                        }
                    },
                    "400": {
                        "description": "失敗",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.ResponseModel": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message_code": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
...
        "member.LoginRequestModel": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "login_ip": {
                    "type": "string"
                },
                "login_url": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "user_agent": {
                    "type": "string"
                }
            }
        },
        "member.LoginResponseModel": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "object",
                    "$ref": "#/definitions/token.TokenModel"
                },
                "user": {
                    "type": "object",
                    "$ref": "#/definitions/member.UcMemberSchema"
                }
            }
        },
...
        "member.UcMemberSchema": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "ag_amount": {
                    "type": "number"
                },
                "amount": {
                    "type": "number"
                },
                "hierarchy": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "last_login_time": {
                    "type": "string"
                },
                "memberLevel": {
                    "type": "integer"
                },
                "member_code": {
                    "description": "會員編號",
                    "type": "string"
                },
                "mobile": {
                    "description": "手機號",
                    "type": "string"
                },
                "real_name": {
                    "description": "真實姓名",
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
...
        "token.TokenModel": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
    }
}